pyproject.toml
[project]
name = "aws-agent"
version = "0.1.0"
description = "Specialized AWS agent built with LangChain/LangGraph for comprehensive AWS operations"
authors = [{name = "AWS Agent Team", email = "agent@example.com"}]
readme = "README.md"
requires-python = ">=3.9"
dependencies = [
    "langchain>=0.1.0",
    "langgraph>=0.0.26",
    "langchain-openai>=0.0.5",
    "langchain-community>=0.0.10",
    "boto3>=1.34.0",
    "pydantic>=2.0.0",
    "pydantic-settings>=2.0.0",
    "python-dotenv>=1.0.0",
    "pyyaml>=6.0",
    "asyncio>=3.4.3",
    "aiofiles>=23.0.0",
    "fastapi>=0.109.0",
    "uvicorn>=0.27.0",
    "websockets>=12.0",
    "httpx>=0.25.0",
    "rich>=13.0.0",
    "click>=8.1.0",
    "keyring>=24.0.0",
    "cryptography>=41.0.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.4.0",
    "pytest-asyncio>=0.21.0",
    "pytest-cov>=4.1.0",
    "black>=23.0.0",
    "ruff>=0.1.0",
    "mypy>=1.5.0",
    "boto3-stubs[essential]>=1.34.0",
]

[project.scripts]
aws-agent = "aws_agent.cli:main"

[build-system]
requires = ["setuptools>=68.0.0", "wheel"]
build-backend = "setuptools.build_meta"

[tool.setuptools]
packages = ["aws_agent"]

[tool.black]
line-length = 100
target-version = ['py39']

[tool.ruff]
line-length = 100
target-version = "py39"
select = ["E", "F", "W", "C90", "I", "N", "UP", "B", "SIM"]

[tool.mypy]
python_version = "3.9"
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = true